backend/pyproject.toml
[tool.poetry]
name = "mailbrew-notes-backend"
version = "0.1.0"
description = ""
authors = ["Rasul Kireev <me@rasulkireev.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.1.3"
djangorestframework = "^3.12.2"
django-cors-headers = "^3.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
